requirements_minimal.txt
# VisionAI-ClipsMaster 最小依赖配置
# 基于代码分析结果，只包含真正使用的核心依赖

# ===== 核心AI和机器学习 =====
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scikit-learn>=1.2.0

# ===== 视频和图像处理 =====
opencv-python>=4.8.0
Pillow>=10.0.0
ffmpeg-python>=0.2.0

# ===== 用户界面 =====
PyQt6>=6.4.0

# ===== 系统和性能监控 =====
psutil>=5.9.0
GPUtil>=1.4.0

# ===== 日志和配置 =====
loguru>=0.7.2
PyYAML>=6.0.1

# ===== 数据处理 =====
pandas>=2.1.0
matplotlib>=3.7.0

# ===== 网络和API =====
requests>=2.31.0

# ===== 文本处理和NLP =====
jieba>=0.42.1
langdetect>=1.0.9

# ===== 进度显示和用户体验 =====
tqdm>=4.65.0
colorama>=0.4.4

# ===== 新发现的缺失依赖 =====
# XML处理 - 用于导出功能
lxml>=4.9.0

# 表格格式化 - 用于报告生成
tabulate>=0.9.0

# 模型管理 - 用于AI模型下载
modelscope>=1.9.0

# ===== 开发和测试（可选） =====
pytest>=7.0.0
pytest-qt>=4.2.0

# ===== 可选增强功能 =====
# 如需完整功能，可取消以下注释：

# 高级NLP功能
# sentence-transformers>=2.2.0
# spacy>=3.6.0
# nltk>=3.8.0

# 音频处理
# librosa>=0.10.0
# soundfile>=0.12.0
# pydub>=0.25.0

# 科学计算增强
# scipy>=1.10.0
# plotly>=5.15.0
# seaborn>=0.12.0

# 模型优化
# accelerate>=0.21.0
# optimum>=1.12.0

# Web界面
# gradio>=3.39.0
# fastapi>=0.100.0
# uvicorn>=0.23.0

# 数据库支持
# sqlalchemy>=2.0.0

# 缓存和性能
# redis>=4.6.0
# joblib>=1.3.0

# 监控和日志
# prometheus-client>=0.17.0

# 云存储
# boto3>=1.28.0
